python_control/requirements.txt
pyserial>=3.5
matplotlib>=3.5.0
numpy>=1.21.0
tkinter  # Usually included with Python
